+++ R/pimage.R
@@ -279,13 +279,15 @@ pimage.matrix <-
     diag = TRUE,
     row_labels = NULL,
     col_labels = NULL,
-    prop = FALSE,
+    prop = isSymmetric(x),
     flip_axes = FALSE,
     reverse_columns = FALSE,
     ...,
     newpage = TRUE,
     pop = TRUE,
     gp = NULL) {
+    force(prop)
+
     x <- as.matrix(x)
 
     # check data