+++ NEWS.md
@@ -1,3 +1,19 @@
+# seriation 1.5.2 (11/26/2023)
+
+## New Features
+- permute for dendrograms gained parameter dist and accepts now seriation 
+  methods.
+- Added method "AOE" for correlation matrices.
+- registry for seriation methods now contains the name of the seriation criterion
+  and a description. seriate_rep now automatically uses the criterion from 
+  the registry.
+- all seriation methods gained parameter rep.
+
+## Bug Fixes
+- optimal.c: use now the correct data type for Rprintf
+- R2E: MacOS with ARM seems to give a different result for eigen. 
+  Removed the deterministic test. 
+
 # seriation 1.5.1 (07/20/2023)
 
 ## New Fearures
@@ -8,10 +24,10 @@
 
 ## Bug Fixes
 - pimage/ggpimage now use zlim correctly to choose the color palette.
-- BEA for matrix is not correctly registered as randomized.
+- BEA for matrix is now correctly registered as randomized.
 - fixed col/row_labels parameter.
-- rev() for seriations based on hclust now also reverse the dendrogram.
-- tests now also accept reverse order for testing deterministic methods.
+- rev() for seriations based on hclust now reverses the dendrogram.
+- tests now also accept reverse orders for testing deterministic methods.
 
 # seriation 1.5.0 (07/19/2023)